more design doc stuff
this is all conceptual level stuff. ideas for how to do things, not any actual code or implementation. I like to put this stuff separatley as a reference guide that can be referred back to as we code stuff. Also, this is what I think we should use to discuss design ideas etc. as texting is a poor way to send large amounts of data.